fix(dto): corrigindo o tipo do campo quantidade Integer -> Long para evitar erro de conversão

+++ src/main/java/com/academico/DataShape/model/dto/responses/QtdVendaDTO.java
@@ -1,5 +1,5 @@
 package com.academico.DataShape.model.dto.responses;
 
-public record QtdVendaDTO(String obra, Integer vendido) {
+public record QtdVendaDTO(String obra, Long vendido) {
     
 }
